requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.1.2
black==24.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.14.0
fsspec==2024.6.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.3
idna==3.7
install==1.3.5
Jinja2==3.1.4
kafka-python-ng==2.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
openai==1.31.1
opencv-python==4.10.0.82
orjson==3.10.3
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
pip==24.0
platformdirs==4.2.2
psycopg==3.1.18
psycopg-binary==3.1.18
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
PyPika==0.48.9
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpc==0.1.0
safetensors==0.4.3
setuptools==69.5.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
sympy==1.12.1
tokenizers==0.19.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.41.2
typer==0.12.3
typing_extensions==4.12.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
validators==0.28.3
watchfiles==0.22.0
websockets==12.0
wheel==0.43.0
